[add] logic coupon

+++ src/main/java/com/ticketsystem/dto/request/CouponCreateRequest.java
@@ -27,6 +27,6 @@ public class CouponCreateRequest {
     @Size(message = "Code length must be >= 5",min = 5)
     String code;
     @Min(message = "discount must be > 0", value = 0)
-    BigDecimal discount;
+    Double discount;
     Integer type;
 }
