+++ src/main/java/com/ticketsystem/dto/request/CouponUpdateRequest.java
@@ -26,6 +26,6 @@ public class CouponUpdateRequest {
     @Size(message = "Code length must be >= 5",min = 5)
     String code;
     @Min(message = "discount must be > 0", value = 0)
-    BigDecimal discount;
+    Double discount;
     Integer type;
 }
